requirements.txt
stomp.py==8.1.0
demucs>=4.0.0
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
httpx==0.24.0
tinytag==1.9.0
pydub==0.25.1
pytest==7.3.1
torch==2.0.1
torchaudio==2.0.2